requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-core>=0.1.20
langsmith>=0.0.70
openai>=1.6.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
